fix managePaper sum

+++ src/app/platform-user/core-controller/dto/receive/managerPaper-create.dto.ts
@@ -15,7 +15,7 @@ export class ManagerPaperCreateDto {
   @IsNotEmpty({ message: 'eventDate is required' })
   @Transform(({ value }) => new Date(value))
   eventDate: Date;
-  @Transform(({ value }) => parseInt(value))
+  @Transform(({ value }) => parseFloat(value))
   @IsNotEmpty({ message: 'sum is required' })
   sum: number;
   @Transform(({ value }) => parseInt(value))
